+++ HRExpert.Database/Organization/DocumentComments.sql
@@ -4,7 +4,7 @@
 	[CreateDate] datetime not null default(getdate()),
 	[DocumentGuid] uniqueidentifier not null,
 	[Message] nvarchar(max),
-	[Author] bigint not null,
+	[Author] int not null,
 	Constraint FK_DocumentComments_Document Foreign key (DocumentGuid) References Documents(Id),
 	Constraint FK_DocumentComments_Author foreign key (Author) References Persons(Id)
 )
