+++ HRExpert.Database/Organization/DocumentTypes.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[DocumentTypes]
 (
-	[Id] bigint NOT NULL PRIMARY KEY identity(1,1),
+	[Id] int NOT NULL PRIMARY KEY identity(1,1),
 	[Name] nvarchar(256) not null
 )
