+++ HRExpert.Database/Organization/Organizations.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[Organizations]
 (
-	[Id] bigint NOT NULL PRIMARY KEY Identity(1,1),
+	[Id] int NOT NULL PRIMARY KEY Identity(1,1),
 	[Code1C] uniqueidentifier,
 	[Name] nvarchar(256)	,
 	[Delete] bit not null default(0)
