+++ HRExpert.Database/Organization/Timesheet/Sicklist/SicklistPaymentPercent.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[SicklistPaymentPercent]
 (
-	[Id] bigint NOT NULL PRIMARY KEY identity(1,1),
+	[Id] int NOT NULL PRIMARY KEY identity(1,1),
 	[Value] decimal(18,2) not null default(0)
 )
