+++ HRExpert.Database/Organization/Timesheet/Sicklist/SicklistTypes.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[SicklistTypes]
 (
-	[Id] bigint NOT NULL PRIMARY KEY identity(1,1),
+	[Id] int NOT NULL PRIMARY KEY identity(1,1),
 	[Code] nvarchar(32),
 	[Name] nvarchar(256) not null
 )
