+++ HRExpert.Database/Organization/Timesheet/TimesheetStatuses.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[TimesheetStatuses]
 (
-	[Id] bigint NOT NULL PRIMARY KEY identity (1,1),
+	[Id] int NOT NULL PRIMARY KEY identity (1,1),
 	[Name] nvarchar(256) not null,
 	[ShortName] nvarchar(32) not null,
